WIP: Add iCloud devices battery status
