+++ .coveragerc
@@ -268,7 +268,7 @@ omit =
     homeassistant/components/hydrawise/*
     homeassistant/components/hyperion/light.py
     homeassistant/components/ialarm/alarm_control_panel.py
-    homeassistant/components/icloud/device_tracker.py
+    #homeassistant/components/icloud/*
     homeassistant/components/idteck_prox/*
     homeassistant/components/ifttt/*
     homeassistant/components/iglo/light.py
