+++ homeassistant/components/icloud/const.py
@@ -1,0 +1,1 @@
+"""Constants for the iCloud component"""
